Added new dataset(s), functions are on top and small LaTeX changes
Added new datasets: ISR is used as another "supporting" dataset. On the main (CPD) dataset, you cannot do any of the statistics, therefore ISR is integrated for statistics. Based on the city (if it matches), district (if it is the same as in CPD) and beat (if it is the same as in CPD), we match the data and can do analysis of age, height and weight of "bad actors". There are other attributes (i.e.: if they are gang members) that can be used in the future.

Functions are added to the top:
mergeData - merges ISR datasets to one
removeNA - removes missing values
dataStatistics - gets data statistic (mean, median, etc...) from passed value
drawPlot - draws distribution plot, bad name, will rename it....
z.test - for performing z test

From LaTeX perspective, small updates have been done, mostly structure of the report.